// Frontend/ClientApp/src/app/models/Company-utilities/company-users.ts
export class CompanyUsers_U {
  isblocked: boolean = true;
  systemmemberships: string = '';
  virtualoperatormemberships: string = '';
  companymemberships: string = '';
  timezone: string = 'America/Bogota';
  languagecode: string = 'es-CO';
  name: string = '';
  email: string = '';
  telephone: string = '';
}